Use CMAKE_INSTALL_LIBDIR instead of hardcode
When packaging an RPM for x86_64, it's uncomforting to find the `libhpdf.so` under $prefix/usr/lib instead of $prefix/usr/lib64 after `make install`, despite setting `CMAKE_INSTALL_LIBDIR` to `lib64`.

The `CMAKE_INSTALL_LIBDIR` variable is even printed out as a *location variable which can be set by the user*. So - use it here!

It'd be nice to also provide a default value for `CMAKE_INSTALL_LIBDIR`, for the case when user doesn't set it explicitly.  
It'd be nice to also handle `CMAKE_INSTALL_INCLUDEDIR` and `CMAKE_INSTALL_BINDIR` in the same fashion.  
I leave this out for now, since it doesn't matter for my use-case (the specfile always sets `CMAKE_INSTALL_*DIR` anyway).

The unrelevant `<foo> DESTINATION` flags are removed, too, in this patch.