fix possible integer overflow
if the constant is set to 1e-20 logVal can take the value -19 leading to prec = 24. Later we set fShift to 10**prec which is too large for a 64bit int. Setting the initial constant to 13 guarantees that fShift is set to a maximum of 1e18 which is just within the INT64 range.