Adapt CMake scripts for WebAssembly compilation (#344)
* Remove useless haru.cmake

This script doesn't seem to be used as the MATH_LIB variable is never used elsewhere and there's another `find_library` call for the math library in the root `CMakeLists.txt`.
Plus, this script incorrectly errors out in case of a compilation for WebAssembly.

* Adapt CMake script for WebAssembly compilation

Don't try to link the Math library when compiling with emscripten, as it's already provided by the system (Compiler throws warnings for missing linked libraries unless compiling to JS emscripten-core/emscripten#2600).
Also added some additional details to CMake summary related with this change.